+++ netbox/circuits/forms/__init__.py
@@ -1,4 +1,4 @@
 from .bulk_edit import *
 from .bulk_import import *
 from .filtersets import *
-from .models import *
+from .model_forms import *
