+++ netbox/tenancy/forms/__init__.py
@@ -1,5 +1,5 @@
 from .forms import *
-from .models import *
+from .model_forms import *
 from .filtersets import *
 from .bulk_edit import *
 from .bulk_import import *
