+++ netbox/virtualization/forms/__init__.py
@@ -1,4 +1,4 @@
-from .models import *
+from .model_forms import *
 from .filtersets import *
 from .object_create import *
 from .bulk_create import *
