+++ netbox/wireless/forms/__init__.py
@@ -1,4 +1,4 @@
-from .models import *
+from .model_forms import *
 from .filtersets import *
 from .bulk_edit import *
 from .bulk_import import *
